[NO LOGIC] [PM-21104] Organize Core.Billing tax code
