+++ src/Core/Billing/Models/PaymentMethod.cs
@@ -1,4 +1,6 @@
-﻿namespace Bit.Core.Billing.Models;
+﻿using Bit.Core.Billing.Tax.Models;
+
+namespace Bit.Core.Billing.Models;
 
 public record PaymentMethod(
     long AccountCredit,
