+++ src/Core/Billing/Models/Sales/CustomerSetup.cs
@@ -1,4 +1,6 @@
-﻿namespace Bit.Core.Billing.Models.Sales;
+﻿using Bit.Core.Billing.Tax.Models;
+
+namespace Bit.Core.Billing.Models.Sales;
 
 #nullable enable
