+++ src/Core/Billing/Services/IPremiumUserBillingService.cs
@@ -1,5 +1,6 @@
 ﻿using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Models.Sales;
+using Bit.Core.Billing.Tax.Models;
 using Bit.Core.Entities;
 
 namespace Bit.Core.Billing.Services;
