+++ src/Core/Billing/Services/IProviderBillingService.cs
@@ -4,6 +4,7 @@ using Bit.Core.Billing.Entities;
 using Bit.Core.Billing.Enums;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Services.Contracts;
+using Bit.Core.Billing.Tax.Models;
 using Bit.Core.Models.Business;
 using Stripe;
