+++ src/Core/Billing/Tax/Requests/PreviewOrganizationInvoiceRequestModel.cs
@@ -1,8 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using Bit.Core.Billing.Enums;
+using Bit.Core.Billing.Models.Api.Requests;
 using Bit.Core.Enums;
 
-namespace Bit.Core.Billing.Models.Api.Requests.Organizations;
+namespace Bit.Core.Billing.Tax.Requests;
 
 public class PreviewOrganizationInvoiceRequestBody
 {
