+++ src/Core/Billing/Tax/Services/IAutomaticTaxFactory.cs
@@ -1,6 +1,7 @@
-﻿using Bit.Core.Billing.Services.Contracts;
+﻿using Bit.Core.Billing.Services;
+using Bit.Core.Billing.Services.Contracts;
 
-namespace Bit.Core.Billing.Services;
+namespace Bit.Core.Billing.Tax.Services;
 
 /// <summary>
 /// Responsible for defining the correct automatic tax strategy for either personal use of business use.
