+++ src/Core/Services/IPaymentService.cs
@@ -2,8 +2,8 @@
 using Bit.Core.AdminConsole.Models.Business;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Models.Api.Requests.Accounts;
-using Bit.Core.Billing.Models.Api.Requests.Organizations;
-using Bit.Core.Billing.Models.Api.Responses;
+using Bit.Core.Billing.Tax.Requests;
+using Bit.Core.Billing.Tax.Responses;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Models.Business;
