+++ test/Core.Test/Billing/Tax/Services/AutomaticTaxFactoryTests.cs
@@ -3,14 +3,14 @@ using Bit.Core.Billing.Enums;
 using Bit.Core.Billing.Models.StaticStore.Plans;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services.Contracts;
-using Bit.Core.Billing.Services.Implementations.AutomaticTax;
+using Bit.Core.Billing.Tax.Services.Implementations;
 using Bit.Core.Entities;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
 using NSubstitute;
 using Xunit;
 
-namespace Bit.Core.Test.Billing.Services.Implementations;
+namespace Bit.Core.Test.Billing.Tax.Services;
 
 [SutProviderCustomize]
 public class AutomaticTaxFactoryTests
